requirements.txt
aiogram==3.4.1
python-dotenv
APScheduler
pandas
openpyxl
SQLAlchemy
aiosqlite
python-dateutil
pytz
xlsxwriter
